SERVER-22541 Refactor RAII locking helpers.
Removes the class 'ScopedTransaction' and moves the responsibility of
abandoning the snapshot to the GlobalLock class. Also renames the
AutoGetCollectionForRead helper to AutoGetCollectionForReadCommand, and
adds a new AutoGetCollectionForRead class. Unlike
AutoGetCollectionForReadCommand, this new class will not update the
namespace on the CurrentOp object, nor will it add an entry to Top.

GitOrigin-RevId: fcd6f14c86ef2ac0be92e4cc4ca8893dcf603b01