SERVER-40477 Fail on type mismatch when producing intent-to-encrypt markings.
Inserts and updates will fail if the value being written to
the server does not comply with the 'bsonType'
specification. Equality comparison to an encrypted field in
reads require deterministic encryption and for exactly one
BSON type to be specified. Mongocryptd will similarly fail
reads when the type of the constant in the query differs
from the type in the schema.

GitOrigin-RevId: 572c8fa1da73f28882b5c864fe9c0f5e05d2d68b