SERVER-77997: [streams] support kafka consumer group ID
Similar to flink and materialize, the offset for the consumer group ID will be
committed to the kafka broker when checkpoint occurs rather than committing in
real time. This means that all messages prior to the committed offset will never
be requested against by the stream processor since it's within the last
checkpoint.

`$source.groupId` on the kafka source operator is optional. If the user does not
supply a consumer group ID, then one will be auto-generated which will be used
for the lifetime of the stream processor.

GitOrigin-RevId: d9b7817e8dc6909e8cd653a13794d4b71bee85eb