SERVER-82814: Support $_internalExpr in clustered coll idx scan
INTERNAL_EXPR_* variants of {EQ, LT, LTE, GT, GTE} are less "precise"
than the non-internal equivalents; that is, they allow additional values
to be returned, and require a second stage of filtering.

When scanning records in a clustered index scan, these variants may
still be re-represented as min/max RecordIDs.