SERVER-66340 Generalise the FeatureCompatibilityVersion lock into a MultiDocumentTransactionsBarrier
This barrier resolves a class of deadlocks involving an uncommitted transaction in the
prepared state and a global lock acquisition request in non-intent mode. Without this
barrier, this would result in a three-way deadlock between (A) a strong global lock
request which depends on (B) a prepared transaction which holds the global lock in intent
mode which depends on (C) the transaction coordinator persisting the commit/abort
decision, which depends on (A), thus establishing a circular dependency.

The FeatureCompatibilityVersion lock was initially introduced to address this issue
during FCV upgrades/downgrades, which acquire the global lock in MODE_S. This
patch generalises this lock into a MultiDocumentTransactionsBarrier to address any
non-intent global lock acquisitions, including fsyncLock.
The MultiDocumentTransactionsBarrier transparently drains prepared transactions
before acquiring the global lock in non-intent mode. It is implicitly acquired as part of a
global lock acquisition by operations processing transaction statements, and by operations
acquiring the global lock in non-intent mode. All other global lock requests skip this
lock. It is acquired in the same mode as the requested global lock mode.

One (benign) behavioural change is that prepared transactions that commit while a global
lock request is enqueued will not be able to acknowledge their write concern to the client
until the global lock is rescinded.