SERVER-81144 Compute QueryShapeHash from QueryShape without full serialization (#17548)
Before this patch, in order to compute SHA256 hash of the query shape, a
full serialization into a BSONObj was necessary, which was later hashed
using SHA256 algorithm. This patch remove the full serialization into
BSONObj and instead relies on hash combining of individual query shape
components. This provides performance boost when computing the hash.