SERVER-85848 Treat $redact as single-doc transform for change stream stages (#18729)
SERVER-85848 Treat $redact as single-doc transform for change stream
stages

The internal stages DSCSEnsureResumeTokenPresent and
DSCSHandleTopologyChange are permitted to swap with single-document
transformation stages during optimization to allow user-specified stages
to be moved closer to the front of the pipeline and to be pushed down to
the shards in a cluster deployment. The $redact stage is allowed in a
change stream pipeline, but it is not internally implemented or
recognised as a single-document transformation; it therefore does not
get moved ahead of the change stream stages mentioned above. This can in
turn decrease the performance for stages after $redact and fail the
$changeStreamSplitLargeEvents stage in a cluster deployment. In this
change we treat the $redact stage similar to single-document
transformation stages during the change stream pipeline optimisation.