SERVER-85337 Replace with custom positional accumulators in $bucketAuto (#18640)
SERVER-85337 Replace with custom positional accumulators in $bucketAuto

$bucketAuto sorts documents internally in order to determine the
boundaries of buckets. That disrupts the order of the documents and
makes the position-aware accumulators like $first to return incorrect
results. To address that, this commit mimics $topN/$bottomN to re-order
the documents for positional accumulators ($first/$firstN/$last/$lastN).
To achieve this, we replace them with the custom one
AccumulatorFirstLastNForBucketAuto.