SERVER-85238 Implement replanning in CachedPlanner for find query (#19144)
SERVER-85238 Implement replanning in CachedPlanner for find query

This commit implements basic replanning algorithm in CachedPlanner. When
a SBE plan is retrieved from cache, it performs SBE trial run phase to
see if replanning is needed. The replanning criteria is to see if the
number of reads (i.e. totalKeysExamined + totalDocsExamined) exceeds the
budget: internalQueryCacheEvictionRatio * decisionReads.