SERVER-86237: Introduce fast path for indexed single equality predicates (#18802)
SERVER-86237: Introduce fast path for indexed single equality predicates

This commit extends the express query executor to support a new class of
queries: single-field indexed equality queries guaranteed to return at
most one document (i.e., the query has a limit 1 or it can use a unique
index on the same field). There are further restrictions on index and
collection types. When a query is eligible to go through the express
path, it is indicated in the explain output with stage "EXPRESS" and in
slow query logs/profiler with planSummary "EXPRESS".