SERVER-87456 Don't store reject: false in query settings (#19729)
The 'reject: false' query setting is a default value, so we can save
some bytes by not storing it.

For the user it makes only sense to specify 'reject: false' when the
value is currently 'true', however, we should accept this setting also
in other cases, because we want 'setQuerySettings' to stay retryable.