SERVER-87377 Add replanReason in MultiPlanStats (#19843)
SERVER-87377 Add replanReason in MultiPlanStats

When Classic Runtime Planner for SBE is enabled and CachedPlanner
decides to replan a query, replanReason is missed when EOF optimization
is used in MultiPlanner. This is because the explainer only access the
stats from the underlying MultiPlanStage and its specific stats does not
have that.

This commit adds an optional field to allow the caller to pass in
replanReason when MultiPlanStage is constructed and hence the reason can
be propagated from CachedPlanner, MultiPlanner to MultiPlanStage.