SERVER-90516 Require magic restore node to run with --replSet and set parameter in targeted tests (#22230)
This PR addresses the issue where we weren't running magic restore nodes
with `--replSet` in our targeted tests. I've now made it a requirement
to pass in the parameter when we run magic restore. Cloud confirmed they
don't support backing up standalones, and we already implicitly require
the node being restored to be a replica set (`--replSetConfig` is a
required restore configuration parameter).

I want to note that the automation credentials test were breaking with
an error `no timestamp provided for an update to a table configured to
always use timestamps`. This was occurring when we tried to do an
untimestamped write to a restored table that had existing timestamped
writes. Turns out we hit this issue in the legacy restore procedure as
well ([HELP-45192](https://jira.mongodb.org/browse/HELP-45192)), so I
needed to add a line of code that skipped the WT timestamp check for
magic restore. This is safe because we're only taking unstable
checkpoints during the restore process, and all timestamps are set
correctly at the end. We also aren't concerned with resumability of the
restore procedure at the moment. The linked HELP ticket has a reference
to the `wiredTigerSkipTableLoggingChecksOnStartup` parameter, but that's
not relevant for magic restore since we're not doing restarts.


[evergreen](https://evergreen.mongodb.com/version/66450586b4a29f0007ba0417)
(all magic restore targeted tests are passing except
`magic_restore_oidc_replica_set.js`, but [that's a known
issue](https://jira.mongodb.org/browse/BF-33184).

---------

Co-authored-by: Alex Neben <alex.neben@mongodb.com>