SERVER-90377 Make config transition hook work for general add and remove shard (#22325)
# Issue

[SERVER-90377](https://jira.mongodb.org/browse/SERVER-90377)

# Description

This patch changes the configshard_transition.py hook (and renames it)
so that it can both transition a config server in and out of dedicated
mode, and can add and remove a random shard (or both, depending on the
suite options). It also adds two concurrency suite definitions - one
that runs add and remove shard in the background continuously, and one
that both adds and removes shards and transitions the config server in
and out of dedicated. It does not actually add these tests to
etc/evergreen yet, and that will be done in SERVER-89866.

# Testing
EVG - runs the new suites to show they run and can successfully complete
some transitions:
https://spruce.mongodb.com/version/664ace004c12f90007a51e38/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC