SERVER-89803 Robustify awaitdata_getmore_cmd.js to filter by specific ns when issuing awaitdata getMore on oplog (#21558)
Before this PR, we issued a find command on the oplog and issued
getMores with awaitdata. When issuing and waiting for getMores, we
expect to stop finding any docs appended to oplog after timeout.
However, background tests can affect this by writing to the oplog.

After this PR, the find command issued is filtered on a specific
namespace. Since we have more control over this namespace, we can ensure
that no doc is written to the oplog after a certain point.


https://spruce.mongodb.com/version/66575c14ccf3b40007cfbc1d/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC