SERVER-90887 Update diff tool to be compatible with thin targets (#22595)
The diff tool relied on configuring which build system to use via
BAZEL_BUILD_ENABLED. After the thin targets change, two build
definitions are no longer permitted.

Since this tool is only intended to be run in CI manually when creating
migration PRs, use git to stash the PR's changes locally to build the
desired targets in SCons.

This relies on the fact that Evergreen stores the PR's changes as staged
files, but also would work if Evergreen stored all of the PR's changes
as a commit.

Diff of environment_buffer (converted to test the diff tool fix out):
> evergreen patch -p mongodb-mongo-master --browse --yes -f --alias
bazel
--param=bazel_scons_diff_targets="mongo/base/libenvironment_buffer.so"

https://spruce.mongodb.com/version/66512d4c736e8700071c55cd/tasks


https://mciuploads.s3.amazonaws.com/mongodb-mongo-master/amazon-linux2-arm64-dynamic-compile/9a87a974a05ec917037908d38451ee8b706131a0/bazel-scons-diff-mongodb_mongo_master_amazon_linux2_arm64_dynamic_compile_compile_dist_test_patch_9a87a974a05ec917037908d38451ee8b706131a0_66512d4c736e8700071c55cd_24_05_25_00_14_09-0.md

Required PB:
https://spruce.mongodb.com/version/6651332fceb9c10007515598/tasks