SERVER-90978: Express: restoreResources() should handle nullptr colle… (#22690)
# Issue
SERVER-90978

# Description
In Express executor, restoreState() calls restoreResources() on the
plan, which passes a CollectionPtr (via a RestoreContext) that can be
null in cases where the executor is using CollectionAcquisitions. We
need to be careful not to dereference this pointer when it is null.

# Testing


[Evergreen](https://evergreen.mongodb.com/version/66575c2c59ccc00007506c1c?redirect_spruce_users=true)

<!--
Available Evergreen Comment Commands
evergreen patch - attempts to create a new PR patch that reruns on each
new commit; this is required to create a PR patch when only manual PR
testing is enabled
evergreen keep-definitions - reuse the tasks from the previous patch in
subsequent patches. Use this after you modify a patch's tasks.
evergreen reset-definitions - reset the patch tasks to the original
definition
evergreen refresh - resyncs PR GitHub checks
evergreen help - comments the above info
 -->