SERVER-90818 Update SBE stage builder to support handle IXSCAN / CLUSTERED_IDXSCAN unions (#22512)
# Issue

SERVER-90818

# Description

A timeseries fuzzer found a query that the SBE stage builder did not
properly support (BF-33317). This problem affects MongoDB both when
`featureFlagSbeFull` is true and when `featureFlagSbeFull` is false.

I came up with a minimized repro, which I posted on the BF ticket.

The root cause of the problem is that the SBE stage builder doesn't
support having unions of IXSCAN stages with "fetched" stages such as the
CLUSTERED_IDXSCAN stage.

This PR updates the SBE stage builder so that it can handle such unions
correctly.

Also, while developing this PR, I noticed that SBE's support for
AND_HASH and AND_SORTED is incomplete. It's pretty easy looking at the
classic engine's implementation to imagine examples where the SBE stage
builder won't produce correct results.

Assuming no query knobs are set, the query planner will never generate
AND_HASH but it could generate AND_SORTED. (This is rare under the
default query knob settings, but it's technically possible.)

To address this, this PR includes changes so that when we're generating
a QSN tree destined for SBE, we never generate a QSN tree that contains
AND_HASH and AND_SORTED (unless 'internalQueryForceIntersectionPlans' is
set to true). We'll still have test coverage for SBE's (incomplete)
implementation of AND_HASH and AND_SORTED with the
"jstests/noPassthrough/and_hash.js" and
"jstests/noPassthrough/and_sorted.js" tests.

# Testing

evergreen patch -p mongodb-mongo-master:
https://spruce.mongodb.com/version/6657bf4e36ccd2000714b233/tasks