SERVER-89599 Improve query shape computation performance for $in match expressions (#22740)
# Issue

SERVER-89599

# Description

The issue: query shape hash is always computed for the Persistent Query
Settings functioning. To compute the query shape all $in values are
examined. When the number of values is large (>100) this step takes
relatively significant percentage of time.
The patch fixes the problem by precomputing the intermediate inputs
required for the query shape computation during initial parsing of the
expression. The intermediate inputs are the first observed value in $in
for each distinct canonical type.

# Testing

- [ ] Unit tests for $in match expression
- [ ] Existing system correctness tests:
https://spruce.mongodb.com/version/664b6b595d04870007fd523d/tasks
- [ ] Performance tests that had regressed (the patch recovers the
performance level):
https://spruce.mongodb.com/task/sys_perf_perf_mongo_perf_standalone.arm.aws.2023_11_query_read_commands_patch_219e31dc1bf4643491480d326d6ec927c438ccf5_663cab73366f9900071c1755_24_05_09_11_02_27?execution=0&sortBy=STATUS&sortDir=ASC

https://spruce.mongodb.com/task/sys_perf_perf_mongo_perf_standalone.intel.aws.2023_11_query_read_commands_patch_219e31dc1bf4643491480d326d6ec927c438ccf5_663cab73366f9900071c1755_24_05_09_11_02_27?execution=0&sortBy=STATUS&sortDir=ASC

---------

Co-authored-by: Mindaugas Malinauskas <mindaugas.malinauskas@mongodb.com>