SERVER-87407 Create targeted jstest for PIT sharded cluster magic restore (#22591)
# Issue

SERVER-87407

# Description

Sorry for the delay, it was blocked by the patch issues with
[SERVER-87406](https://jira.mongodb.org/browse/SERVER-87406).
This is the PIT version of it.

It is pretty similar except for calculating
pointInTimeTimestamp/maxlastOplogEntryTs and the expectation around the
documents inserted after the backup cursor.
There is a change in assertStableCheckpointIsCorrectAfterRestore as some
of the shards might not have the global pointInTimeTimestamp as the
stable timestamp, the correct check is that the stable timestamp is the
last entry in the node own oplog.

# Testing


[Patch](https://spruce.mongodb.com/version/66579745f6279d000714d7f1/tasks)