SERVER-90236 Improve pre-image truncate marker generation observability (#22548)
# Issue
[SERVER-90236](https://jira.mongodb.org/browse/SERVER-90236) Improve
pre-image truncate marker generation observability

# Description
Introduces log statements (1) when pre-image markers are installed in
the _tenantMap and (2) when pre-image markers complete initialization.
New and existing logs now include a "preImagesCollection" attribute with
the UUID of the pre-images collection.

This is the base work for a series of 8.0 backport candidates
([SERVER-90301](https://jira.mongodb.org/browse/SERVER-90301),
[SERVER-90305](https://jira.mongodb.org/browse/SERVER-90305),
[SERVER-90705](https://jira.mongodb.org/browse/SERVER-90705))