SERVER-91021 Handle TransactionParticipantFailedUnyield in internal transactions concurrency tests (#22790)
# Issue

SERVER-91021

# Description

This patch makes it so that concurrency tests which test internal
transactions directly against shards handle the
TransactionParticipantFailedUnyield error - in normal execution mongos
should catch this error and rethrow the "original error" in the extra
info, so we do not expect this to ever bubble up to the client. In tests
where we talk directly to shards, the test infrastructure should do this
handling instead to avoid unnecessarily failing the test.

# Testing

https://spruce.mongodb.com/version/6658e8ee270606000722ac4a/tasks?page=0&sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC&taskName=concurre