SERVER-88502 Add a new nightly build variant to run replica set endpoint suites (#22804)
# Issue

[SERVER-88502](https://jira.mongodb.org/browse/SERVER-88502)

# Description

This is part 1 of SERVER-88502. Here is a summary of the changes:
- Add a new nightly variant `"* Shared Library Enterprise RHEL 8.0
(Replica Set Endpoint all feature flags)"` that runs all suites with the
`"replica_set_endpoint"` tag.
- Remove the "default" tag from all replica set endpoint tasks except
the suites below. That is, they are the only suites that will continue
to run on the (default) variants other than the new build variant above.
The motivation is that currently there are 18 replica set endpoint
suites and they have been running on EVG waterfall for more than 2
months. For the purpose of saving costs, I think we can now start
running them less. In addition to the suites below, there are still
targeted tests in the noPassthrough suite and unit tests that run on the
default variants so if someone breaks the replica set endpoint code,
they would catch it when they run on EVG patch build.
    - replica_sets_jscore_passthrough_with_replica_set_endpoint
    - sharding_jscore_passthrough_with_replica_set_endpoint
    - concurrency_replication_with_replica_set_endpoint
    - concurrency_replication_multi_stmt_txn_with_replica_set_endpoint

# Testing

The EVG patch build below shows the suites that run in each relevant
build variant.

[EVG](https://spruce.mongodb.com/version/6659d7851dc3d50007151157/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)