SERVER-91066 Set migrationCommitVersionError failpoint on all configsvr nodes in migration_coordinator_failover_include.js (#22827)
# Issue

[SERVER-91066](https://jira.mongodb.org/browse/SERVER-91066)

# Description

Some of the test cases in migration_coordinator_abort_failover.js, which
calls functions in migration_coordinator_failover_include.js, set a
failpoint on the configsvr that should force a moveChunk to fail. In
tests that run with a config shard, it's possible that the config shard
is the donor of the moveChunk which the test later forces to failover.
Failpoints aren't persisted across replica set nodes, so the new primary
won't have the failpoint set and the moveChunk succeeds when we don't
expect it to. This patch changes the test to set the failpoint on every
node in the configsvr, so if it is a config shard that acts as the
donor, the failpoint will be set on the new primary as well.

# Testing


https://spruce.mongodb.com/version/6659edeb4d6f8d000745a33d/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC