SERVER-89542: Update pooled LDAP connection timeout at runtime (#22093)
# Ticket
SERVER-89542

# Description
Previously, `ldap_runtime_configuration.js` waited on a failpoint that
stalled LDAP binds twice - one triggered by an auth attempt before an
LDAP param was changed from a good to bad value and one after. After
turning off the failpoint, the first attempt was expected to succeed
while the second was expected to fail.

Unfortunately, the second failpoint wait was not actually happening
since it had already been waited on once, so it simply returned
immediately, causing race conditions between the failpoint being turned
off and the second auth attempt.

Additionally, the failpoint was not very useful for testing the "bad
parameter" case for `ldapTimeoutMS`. The "bad" value of `ldapTimeoutMS`
was a very small value (50 ms), that is expected to cause more timeouts
than the default value (10 seconds). However, an endless hang due to the
failpoint doesn't really test this since they cause failures under both
circumstances.

Now, the test does the following:
- A client attempts authenticating as an LDAP user, and the server
stalls while binding to the
 LDAP server.
- A different client changes LDAP configuration to an invalid value that
should result in LDAP
 auth failures.
- The ongoing authentication attempt should continue using the old
configuration and eventually
 succeed.
- Subsequent authentication attempts should fail due to the now-invalid
LDAP configuration.
- Changing the LDAP configuration back to the original value should
allow auth to start working
 normally again.

Notably, it only uses the failpoint once in order to prove that auth
attempts that start an LDAP bind before a config option is changed are
unaffected by config change. Auth attempts after the config change are
expected to use the new values, which have been selected to cause
failure.
The timeout test specifically sets up a mock LDAP server with a 5 second
delay. This should not cause timeouts in the old timeout value (10
seconds) but should with the new value (50 ms).

While doing this, I found a bug that prevented the LDAP timeout from
being promptly propagated to pooled LDAP connections. The timeout only
got set on connections during setup, meaning that already-running pooled
connections would not update. Now, I ensure that the timeout gets
updated during connection refresh. I also force all pooled LDAP
connections to get dropped when the timeout is updated so that the new
timeouts are used as soon as new connections are made.

Additionally, I've made a small change to the `ldapServers` parameter so
that it returns both the LDAP server URI and port number. I may modify
this so that it's only done if the port is explicitly specified in the
set parameter config or in test-mode.

# Testing
EVG:
https://spruce.mongodb.com/version/663d192da993340007220f5c/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC