SERVER-89948 Make sure acceptable errors are recorded when applying CRUD ops. (#22766)
# Issue

SERVER-89948

# Description

Previously if applying a CRUD op results in an acceptable error, it's
silently ignored and we still log "Applied OP...", which is confusing
and makes it hard to debug issues. This patch makes sure the error is
logged while still leaving the "Applied OP..." message.

Note that this is just a temporary fix, there are also similar issues in
higher levels than what this patch fixes, see bullet point 2 and 3 in
the ticket description. A complete fix/refactoring will be deferred to
SERVER-78834.

# Testing

https://spruce.mongodb.com/version/6658fc29f8caa200077a9e28

<!--
Available Evergreen Comment Commands
evergreen patch - attempts to create a new PR patch that reruns on each
new commit; this is required to create a PR patch when only manual PR
testing is enabled
evergreen keep-definitions - reuse the tasks from the previous patch in
subsequent patches. Use this after you modify a patch's tasks.
evergreen reset-definitions - reset the patch tasks to the original
definition
evergreen refresh - resyncs PR GitHub checks
evergreen help - comments the above info
 -->