SERVER-90157 Never coordinate multi upserts (#22829)
# Issue

[SERVER-90157](https://jira.mongodb.org/browse/SERVER-90157)

# Description

Aside from never coordinating multi upserts, this also refactors the
associated concurrency test to run upserts (which required tracking
individual documents instead of a single expected value for all
counters).

Additionally, the sessions collection no longer does multi deletes when
removing a session. This was making reaping very slow when the cluster
parameter was on. The collection is sharded on _id and uses an exact _id
query when doing deletes, so there was no reason for this to be a multi
delete in the first place.

# Testing


https://spruce.mongodb.com/version/6659312d6b82970007a6d19e/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

None of the failures appear to be related; none of the failing suites
enable the cluster parameter either.