SERVER-91124 Increase defaultMaxTimeMS value to avoid zero remaining time sent to shards in jstest (#22916)
# Issue
[SERVER-91124](https://jira.mongodb.org/browse/SERVER-91124)

# Description
We use maxTimeNeverTimeOut failpoint on mongos to avoid timing out
before the command reaching the shards. But if the default timeout value
is small, it could be used up on mongos and the value zero could be sent
which means "do not time out". This can lead to test cases which are
expected to fail to pass. We should give it enough time on the router to
avoid this scenario.

# Testing

https://spruce.mongodb.com/version/665ef52b2192320007b5cde8/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC