SERVER-90378 : Cache reference to _id index in IndexCatalogImpl (#22810)
# Issue

SERVER-90378 

# Description

When determining if a query is eligible for the express path, we check
for the presence of an _id index, which requires going to the catalog
and iterating through the index list. Later, we retrieve the _id index
again, this time to actually execute the query. To avoid this double
iteration, we can instead cache the _id index descriptor on the
IndexCatalogImpl. This benefits queries that do not use the express path
as well.

Note the similarity to the approach in
https://github.com/10gen/mongo/pull/22130/files. The difference is that
the cached _id index is set on index creation (in `createIndexEntry()`
and `indexBuildSuccess()`) rather than within `findIdIndex()`. Unlike
`findIdIndex()`, these methods should happen under a collection X lock,
which will resolve the race seen when multiple oplog application threads
using the same IndexCatalogImpl try to set the cached pointer at the
same time. Similarly, the other method that writes to the pointer,
`dropIndexEntry()`, should happen under a collection X lock.

# Testing


https://spruce.mongodb.com/version/6659ec474d6f8d0007459e45/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC


https://spruce.mongodb.com/version/665da15edcf05c0007cfb791/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC


https://performance-analyzer.server-tig.prod.corp.mongodb.com/perf-analyzer-viz/?comparison_id=ccc33549-0ce7-4466-8d3d-97bd8154af31&selected_tab=scatter-plots&percent_filter=0%7C%7C100&z_filter=0%7C%7C10


https://performance-analyzer.server-tig.prod.corp.mongodb.com/perf-analyzer-viz/?comparison_id=7589f9f0-83e2-4d71-b7f2-6b025f40a8b5&selected_tab=data-table&context_filter=High-Value+Workloads&percent_filter=0%7C%7C100&z_filter=0%7C%7C10

Colin checked instruction & allocation counts locally for find-by-_id
and saw that this approach performed better when compared to the
baseline and to the approach from
https://github.com/10gen/mongo/pull/22742.