SERVER-91182: Make schedule_global_auto_tasks patch_only. (#22965)
schedule_global_auto_tasks, schedule_variant_auto_tasks, and
schedule_patch_auto_tasks should be set to patch_only, so that they
don't conflict with generate_all_variant_auto_tasks.

# Issue

SERVER-91182

# Description

This fixes BF-32196, which is caused when all tasks on a waterfall build
are run,

# Testing

This is hard to test since it only impacts mainline builds, but it I
validated that the yaml parses and those commands already don't run on
waterfall builds unless they are manually run.