SERVER-89697 remove function sanitizer coverage from wt bin (#22844)
Since wt is built in C mode, we need to disable the function sanitizer
(part of ubsan) in wt and all third party dependencies to avoid missing
symbol errors when dependent libraries are built with ubsan enabled.

Function sanitizer coverage isn't the most important especially in C++,
from the docs it checks for "Indirect call of a function through a
function pointer of the wrong type."

It's a little more important in C, where raw function pointers are more
common. Because of this, it seems bad to disable it in wiredtiger.
However, wiredtiger is running ubsan in its own project, which will give
us coverage of it:
https://github.com/wiredtiger/wiredtiger/blob/7429bc4a9aa94baa7ff707ae80765d08026fcd77/cmake/configs/modes.cmake#L137

This currently is causing (many) failures in
https://jira.mongodb.org/browse/BF-32299

This change is dependent on https://github.com/10gen/mongo/pull/22842 to
actually fix anything.

PBs:

Required + ubsan/aubsan variants:
https://spruce.mongodb.com/version/665f49a724a7ed000773c85d/tasks
Bazel: https://spruce.mongodb.com/version/665f49871c2e530007070c88/tasks