SERVER-90828:  Return EOF solution for trivially false filter ixscans (#22619)
SERVER-90828

Queries such as

```
{a:100, b:{"$in":[]}}
```

are trivially false; nothing will match the empty in.

SERVER-70477 made use of this, and such filters now result in an
`AlwaysFalseMatchExpression` (filter becomes `{$alwaysFalse:1}`).

SERVER-81863 then ensured that collection scans with $alwaysFalse filter
would instead lead to an EOF solution.
However, for ixscans, an $alwaysFalse filter could still require a scan
and fetch, only to discard everything.

This actually made some queries _slower_, as in BF-32974 the index scan
could no longer be bounded by other fields (e.g., in the above example,
the filter on `a` can lead to MinKey=MaxKey=100), requiring a fetch and
filter.

To resolve, this regression make ixscan behaviour consistent by
returning an EOF solution if the filter is trivially false; it cannot
match anything.


[evergreen](https://spruce.mongodb.com/version/6655dd787050f30007113738),
[sys-perf](https://performance-analyzer.server-tig.prod.corp.mongodb.com/perf-analyzer-viz/?comparison_id=df21f2e1-1e66-4cab-8681-e24d897b342f).