SERVER-91173 Retry on errors from targeting removed shard in add_remove_shards_killop.js (#22952)
EVG:
https://spruce.mongodb.com/version/665f6050c4406c0007fcbec8/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

It's a known issue that removing a shard will interrupt anything using
its ReplicaSetMonitor when the RSM is removed. The $currentOp this test
uses can hit that scenario when the RS endpoint is on, since it will
target all shards even though it's sent directly to a config server
node. The returned error is ShutdownInProgress, which is a "retryable"
error, so I reused the helper to retry on such errors already in this
test.