SERVER-91034 Add assert for collection existence in validate_checkBSONConformance.js (#22870)
# Issue

SERVER-91034

# Description

BF-32263 saw the "test.validate_checkBSONConformance" collection
disappearing before it could be validated, or perhaps failing to be
created. We don't have any further data to investigate, and the issue
has not recurred for several weeks, but if it happens again an
additional assert here will provide some more detail as to the timing of
the issue. Also moving the tests that require the collection up front
will reduce the likelihood of future recurrence.

# Testing

https://spruce.mongodb.com/version/6658e41012fd160007952bdf