SERVER-87285 Make find command batch size configurable (#22533)
# Issue
SERVER-87285

# Description

This PR makes the find command batch size configurable by adding the
query knob 'internalQueryFindCommandBatchSize.' Previously, this value
was hardcoded at 101 so getMore commands were not exercised as often as
they could be.

# Testing

Setting the query knob is tested in query_knobs_validation.js. Existing
find tests were modified to account for in the
internalQueryFindCommandBatchSize knob when batch size was used. The
value was also added to the config fuzzer for both mongos and mongod.

[Patch
🌲](https://spruce.mongodb.com/version/66579cf1e50eb700073af2ba/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

Note that the following burn_in tests also fail/time out on current
master:
- change streams or replica set suites with periodic kill secondaries
for find9.js
- sharding_uninitialized_fcv_jscore_passthrough for find9.js and
exhaust.js on Enterprise RHEL 8.0