SERVER-74657 Address StorEx threads that are marked unkillable (#22937)
# Issue

SERVER-74657

# Description

<!-- Please include a summary of the changes and the related issue. Be
sure to include relevant motivation and context. -->
In [SERVER-70127](https://jira.mongodb.org/browse/SERVER-70127), we
changed the default behavior of a thread to be killable by default. We
now marked all threads that were not marked as killable before to be
unkillable, but it could be there are threads not necessary to be
unkillable.
This ticket is to revisit 10 threads assigned to StorEx team.

The criteria is that if a thread is primary/secondary agnostic and
doesn't do WT write, the thread is kept as unkillable, 8 threads pass
this criteria. For the 2 threads changed to killable, the exception is
handled safely.

# Testing

- [ ]
https://spruce.mongodb.com/version/665f37ea81c7ff0007043795/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

<!--
Available Evergreen Comment Commands
evergreen patch - attempts to create a new PR patch that reruns on each
new commit; this is required to create a PR patch when only manual PR
testing is enabled
evergreen keep-definitions - reuse the tasks from the previous patch in
subsequent patches. Use this after you modify a patch's tasks.
evergreen reset-definitions - reset the patch tasks to the original
definition
evergreen refresh - resyncs PR GitHub checks
evergreen help - comments the above info
 -->