SERVER-91230 Change SBE SubPlanner to write 'reads' instead of 'works' when planning entire query (#23031)
# Issue

https://jira.mongodb.org/browse/BF-33563
https://jira.mongodb.org/browse/SERVER-91230

# Description

There are two paths where the SubPlan stage can write to the cache:
1 - Planning individual branches. The usual case.
2 - Planning the whole query. This is used when we bail out of case (1)

In the SBE SubPlanner code, the callback for (2) writes a "works" value
instead of a "reads" value. This is problematic, because it means when
recovering the cache entry, we do not have a "reads" value to compare
with. The result is that a tassert() is tripped
[here](https://github.com/10gen/mongo/blob/bdaea545b0dab4763fe82dceeed72d09f5576615/src/mongo/db/query/plan_cache.h#L116-L117)
when reading from the cache. This was discovered on the trySbeEngine
variant, which uses SBE for all queries possible with the classic plan
cache.

The fix is to make the callback provided by the SBE SubPlanner write a
"reads" value instead of a "works" value. This required some plumbing,
but conceptually it's not a huge change.

# Testing

https://spruce.mongodb.com/version/6660add3a986420007e17705/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC