SERVER-91199 Allow more time for mongod on slow builds to catch up the min snapshot read time. (#23021)
# Issue

SERVER-91199

# Description

Consider the following sequence of events:

| T | Event |
| :---: | :--- |
| T0 | Insert document D |
| T1 | <I’ll come back to this> |
| T2 | Update document D => D’ |
| T3 | Wait minSnapshotHistoryWindowSecs + margin |
| T4 | Unrelated w-majority write to trigger history cleanup |
| T5 | Read with afterClusterTime=T0 |

We expect to see the updated doc D’ from T2 because anything before T2
should be removed from history, but on slow builds, we sometimes see the
doc D from T0 with atClusterTime=T1. This behavior is correct since T1 >
T0, but since we explicitly wait at T3, we expect mongod to catch up
eventually such that the read at T5 sees D'.

In this PR, we add some extra delay if the read at T5 is served with a
clusterTime between T1 and T2.

# Testing

- [x] [Patch
1](https://spruce.mongodb.com/version/66608101e152050007fa6551/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

[Patch
2](https://spruce.mongodb.com/version/6660a3ebd9807d00071a363a/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

[Patch
3](https://spruce.mongodb.com/version/6660a41462a1720007bbdb4e/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)