SERVER-90544 Improve performance of cpu thread timer (#22644)
# Issue

SERVER-90544

# Description

Improves performance of cpu thread timer, by removing need for malloc.
Pausing the time progression on OperationCPUTimers when thread is
detached level eliminates the need to register and propagate the
deteach/attach operations to all timers.

The running timer count is tracked in order to avoid clock_gettime calls
when they would be unnecessary, that is when no timers are running.

# Testing

This reduces the lifecycle cost of cpu timer by around 200-500ns when
measured as part of find command execution.
Based on the sep ping benchmark, it reduces number of instructions by
300, which is ~1.3%


https://spruce.mongodb.com/version/6661cafc5d357700070de44e/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

has the same failures as master run:

https://spruce.mongodb.com/version/6661cad24c81140007a52988/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC