SERVER-90846 Make moveCollection join-able again (#22786)
# Issue

<!-- Jira issue key and link. Entering a SERVER issue will automatically
create a link.-->
SERVER-90846

# Description

<!-- Please include a summary of the changes and the related issue. Be
sure to include relevant motivation and context. -->
Allow moveCollection to join when resharding has already started when
the new command is issued. We are doing this by handling the LockBusy
error to ensure that we don't have a scenario in which we check for
ongoing reshardings and then the result of that check changes.

# Testing

-
[evg](https://spruce.mongodb.com/version/66599f0f6caf4f000734d3ad/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
<!-- Please describe the tests you ran to verify your changes and
include any links to patches. Consider using `evergreen patch` described
below.-->

<!--
Available Evergreen Comment Commands
evergreen patch - attempts to create a new PR patch that reruns on each
new commit; this is required to create a PR patch when only manual PR
testing is enabled
evergreen keep-definitions - reuse the tasks from the previous patch in
subsequent patches. Use this after you modify a patch's tasks.
evergreen reset-definitions - reset the patch tasks to the original
definition
evergreen refresh - resyncs PR GitHub checks
evergreen help - comments the above info
 -->