SERVER-91322 resmoke multiversion-config: set evergreen config options (#23198)
# Issue

https://jira.mongodb.org/browse/SERVER-91322

# Description

Currently in resmoke our multiversion-config subcommand does not set the
evergreen config options in resmoke. However, we have different behavior
for grabbing the releases.yml file when resmoke is being ran in
evergreen or not. We try to determine this by checking for the evergreen
task id here
https://github.com/10gen/mongo/blob/49f343c18082a1aad4d90889b8eebf8a51b7f251/buildscripts/resmokelib/multiversionconstants.py#L80

This check will always fail because the config option is not actually
set.

This is causing old branches to use outdated releases files and run
multiversion testing against eol versions.

# Testing

Full PB: https://spruce.mongodb.com/version/66634d7a2679440007bc041c
v7.3 PB showing that it is not testing against v7.2 anymore:
https://spruce.mongodb.com/version/6663560e9129d90007ee9838