SERVER-90121 Add shard renames to targeted sharded cluster magic restore tests (#22911)
# Issue

SERVER-90121

# Description

Added those checks in the existing PIT sharded test. Concurrent DDL
operations will be covered by SERVER-87427 and SERVER-87429 as discussed
with Ali.

I found a bug in the technical document (and Automated Restores
Architecture document) as it is `host` not `hosts` in config.shards.

Some of the collection don't exist (at least if there is no concurrent
DDL) so I added NamespaceNotFound exceptions to the fasserts.

# Testing

[Patch](https://spruce.mongodb.com/version/665e56be057092000790937c/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)