SERVER-91254 Convert to decimal before comparison in window (#23185)
# Issue

SERVER-91254

# Description

In window, when computing the threshold we convert the numeric values to
decimals and then add them together. The result of the addition is the
threshold we use to decide whether a document is inside the window. If
the values used to compute the threshold are double we might lose
precision when converting them to decimals. In a recent bug, this lose
in precision had as a consequence the document that was used to compute
the threshold to be found outside the window when compared against the
threshold. The reason was that the value of the document was double and
even though we convert it to decimal when computing the threshold, we
keep it as a double when we compare it against this threshold.

Since decimals and doubles are not exact values, with this change, I am
converting the numeric values to decimals before comparing them against
the threshold. This way, if there is any loss in precision at least the
same value is used when computing the threshold and when comparing
against it.

# Testing


https://spruce.mongodb.com/version/6666eaf7a6569b000663aa31/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

---------

Co-authored-by: Foteini Alvanaki <foteini.alvanaki@mongodb.com>