SERVER-90774: Track number of documents seen so far in idLookup step (#23061)
# Issue

https://jira.mongodb.org/browse/SERVER-90774

# Description

Adds the `_docsSeenByIdLookup` to the `SharedSearchState` class that
counts the number of documents seen so far in by the idLookup stage of
the query plan. Then a `getDocsLookupByIdSuccessRate` is added to
retrieve the success rate of finding docs in the idLookup stage. This
function will later be used for batch size tuning.

# Testing

Added unit tests for the `getDocsLookupByIdSuccessRate` function. Ran
resmoke integration core test suite.

[Updated
🌲](https://spruce.mongodb.com/version/666361e25dd0af0007b596db/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)