SERVER-80623 Support --disable-warnings-as-error on Linux/Windows in Bazel (#23047)
This PR enables treating warnings as errors by default on all Bazel
compilations on Linux and Windows. This is necessary to convert targets
developed by Mongo (distinct from third party targets) because
developers expect C++ compiler warnings to fail builds in Evergreen and
on local workstations.

Many C++ compiler warnings indicate critical correctness issues that
have very significant downstream effects. For example, an issue flagged
by a warning may not necessarily result in an immediate crash, but
instead may result in memory corruption bug one in a hundred thousand
executions that may corrupt user data. Warnings as errors can be
considered a type of sanitization the compiler does by default.

Certain C++ flags will trigger a warning when passed to the C compiler
(ex. clang instead of clang++). This is because clang (for example) does
not have all flags that clang++ (for example) has since they're
compilers for separate (but similar) languages.

If we enabled warnings-as-errors in these cases, they would result in a
build failure. Bazel makes this situation worse because it does not
allow the build system to pass flags only when compiling C code. Any
flags added by default will be applied in both C and C++ compilation.

To get around this, this PR utilizes the custom Linux toolchain config
we have setup to create
[features](https://bazel.build/docs/cc-toolchain-config-reference#features)
with flags that are only applied to C++ compilation steps. We currently
do not have a custom toolchain for Windows or MacOS, which is why we
have to disable warnings-as-errors on MacOS for now. Windows does not
complain about this issue, so we can leave it enabled on Windows.
Warnings-as-errors isn't nearly as important on MacOS as it is on Linux,
since Linux already runs the checks on Clang. We also don't have
customers using MacOS in production.

https://jira.mongodb.org/browse/SERVER-90183 tracks the work to enable
custom toolchain config on MacOS and in turn enable warnings-as-errors
on MacOS.

PB: https://spruce.mongodb.com/version/665fd517c38ba80007580d45/tasks

---------

Co-authored-by: Udita Bose <udita.bose@mongodb.com>